+++ CHANGELOG.md
@@ -8,7 +8,11 @@ to [Common Changelog](https://common-changelog.org)
 
 ### Fixed
 
+ - Fix KEGG module completeness calculation script. ([#63](https://github.com/metagenlab/zshoman/pull/63)) (Niklaus Johner)
+
 ### Changed
+
+- Parallelize and speed-up output merging script. ([#60](https://github.com/metagenlab/zshoman/pull/60) and [#62](https://github.com/metagenlab/zshoman/pull/62)) (Niklaus Johner)
 
 ### Added